Move components to reduce hierarchy, add example tests we can remove later
