+++ plugin-hrm-form/src/Views/CallTypeButtons.js
@@ -1,8 +1,8 @@
 import React from 'react';
-import { Container, StyledButton, StyledFinishButton } from '../../Styles/HrmStyles';
+import { Container, StyledButton, StyledFinishButton } from '../Styles/HrmStyles';
 import { withTaskContext } from "@twilio/flex-ui";
-import { callTypes } from '../../states/DomainConstants';
-import { isStandAloneCallType } from '../../states/ValidationRules';
+import { callTypes } from '../states/DomainConstants';
+import { isStandAloneCallType } from '../states/ValidationRules';
 
 const CallTypeButtons = (props) => {
   // TODO(nick): Figure out how to get these to be separate
