+++ plugin-hrm-form/src/Views/HrmForm.js
@@ -2,7 +2,7 @@ import React from 'react';
 import { withTaskContext } from "@twilio/flex-ui";
 import CallTypeButtons from './CallTypeButtons';
 import TabbedForms from './TabbedForms';
-import { isStandAloneCallType } from '../../states/ValidationRules';
+import { isStandAloneCallType } from '../states/ValidationRules';
 
 // It is recommended to keep components stateless and use redux for managing states
 const HrmForm = (props) => {
